:truck: Move Routes -> routes
