+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {BrowserRouter as Router} from 'react-router-dom';
-import Routes from './Routes';
+import Routes from './routes';
 import {ApolloProvider} from '@apollo/react-common';
 import {client} from './state/client';
 import {AnalyticsTrackingProvider} from './analyticsTracking';
